+++ src/js/validator/emailAddress.js
@@ -4,6 +4,12 @@
     });
 
     $.fn.bootstrapValidator.validators.emailAddress = {
+        html5Attributes: {
+            message: 'message',
+            multiple: 'multiple',
+            separator: 'separator'
+        },
+
         enableByHtml5: function($field) {
             return ('email' === $field.attr('type'));
         },
@@ -14,8 +20,8 @@
          * @param {BootstrapValidator} validator Validate plugin instance
          * @param {jQuery} $field Field element
          * @param {Object} [options]
-         * - allowMultiple: Allow multiple email addresses, separated by a comma or semicolon; default is false.
-         * - separatorRegex: Regex for character or characters expected as separator between addresses; default is comma /[,;]/, i.e. comma or semicolon.
+         * - multiple: Allow multiple email addresses, separated by a comma or semicolon; default is false.
+         * - separator: Regex for character or characters expected as separator between addresses; default is comma /[,;]/, i.e. comma or semicolon.
          * @returns {Boolean}
          */
         validate: function(validator, $field, options) {
@@ -26,61 +32,56 @@
 
             // Email address regular expression
             // http://stackoverflow.com/questions/46155/validate-email-address-in-javascript
-            var emailRegExp = /^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-
-            var allowMultiple = options.allowMultiple === true || options.allowMultiple === 'true';
+            var emailRegExp   = /^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/,
+                allowMultiple = options.multiple === true || options.multiple === 'true';
 
             if (allowMultiple) {
-                var separatorRegex = options.separatorRegex || /[,;]/;
-                var areValid = true;
-                var addresses = splitEmailAddresses(value, separatorRegex);
+                var separator = options.separator || /[,;]/,
+                    addresses = this._splitEmailAddresses(value, separator);
 
                 for (var i = 0; i < addresses.length; i++) {
-                    areValid = areValid && emailRegExp.test(addresses[i]);
+                    if (!emailRegExp.test(addresses[i])) {
+                        return false;
+                    }
                 }
 
-                return areValid;
-            }
-            else {
+                return true;
+            } else {
                 return emailRegExp.test(value);
             }
+        },
+
+        _splitEmailAddresses: function(emailAddresses, separator) {
+            var quotedFragments     = emailAddresses.split(/"/),
+                quotedFragmentCount = quotedFragments.length,
+                emailAddressArray   = [],
+                nextEmailAddress    = '';
+
+            for (var i = 0; i < quotedFragmentCount; i++) {
+                if (i % 2 === 0) {
+                    var splitEmailAddressFragments     = quotedFragments[i].split(separator),
+                        splitEmailAddressFragmentCount = splitEmailAddressFragments.length;
+
+                    if (splitEmailAddressFragmentCount === 1) {
+                        nextEmailAddress += splitEmailAddressFragments[0];
+                    } else {
+                        emailAddressArray.push(nextEmailAddress + splitEmailAddressFragments[0]);
+
+                        for (var j = 1; j < splitEmailAddressFragmentCount - 1; j++) {
+                            emailAddressArray.push(splitEmailAddressFragments[j]);
+                        }
+                        nextEmailAddress = splitEmailAddressFragments[splitEmailAddressFragmentCount - 1];
+                    }
+                } else {
+                    nextEmailAddress += '"' + quotedFragments[i];
+                    if (i < quotedFragmentCount - 1) {
+                        nextEmailAddress += '"';
+                    }
+                }
+            }
+
+            emailAddressArray.push(nextEmailAddress);
+            return emailAddressArray;
         }
     };
-
-    function splitEmailAddresses(emailAddresses, separatorRegex) {
-        var quotedFragments = emailAddresses.split(/"/),
-            quotedFragmentCount = quotedFragments.length,
-            emailAddressArray = [],
-            nextEmailAddress = "";
-
-        for (var i = 0; i < quotedFragmentCount; i++) {
-            if (i % 2 === 0) {
-                var splitEmailAddressFragments = quotedFragments[i].split(separatorRegex);
-                var splitEmailAddressFragmentCount = splitEmailAddressFragments.length;
-
-                if (splitEmailAddressFragmentCount === 1){
-                    nextEmailAddress += splitEmailAddressFragments[0];
-                }
-                else {
-                    emailAddressArray.push(nextEmailAddress + splitEmailAddressFragments[0]);
-
-                    for (var j = 1; j < splitEmailAddressFragmentCount - 1; j++) {
-                        emailAddressArray.push(splitEmailAddressFragments[j]);
-                    }
-
-                    nextEmailAddress = splitEmailAddressFragments[splitEmailAddressFragmentCount - 1];
-                }
-            }
-            else {
-                nextEmailAddress += '"' + quotedFragments[i];
-
-                if (i < quotedFragmentCount - 1)
-                    nextEmailAddress += '"';
-            }
-        }
-
-        emailAddressArray.push(nextEmailAddress);
-
-        return emailAddressArray;
-    }
 }(window.jQuery));